added feature to generate directory structure from timestamp and transfer files to them
This commit adds feature to transfer files from a directory to directories organised as
year/month/day/* format. First, the files in the specified directory are listed. Then,
timestamp is obtained from exif or last modified date or created date. New directories are
generated from this timestamp and files are transfered.